// src/types/settings.ts
export type CardSettings ={
    backgroundColor: string;
    padding: number;
    borderRadius: number;
    width: number;
    height: number;
    titleFontSize: number;
    descriptionFontSize: number;
    tagFontSize: number;
    textColor: string;
    shadow: boolean;
    imageOpacity: number;
    backgroundSize: string;
    backgroundPosition: string;
}